Create new methods that build the request routes
create ``sermonsApiRoute`` ``speakersApiRoute`` ``eventsApiRoute``
``languagesApiRoute`` and ``buildRequestRoute`` methods so that the API
user is not forced to make a request with the ``requestData`` method
provided in the class. This will allow the user to create the request
route via this class and then make the request using whatever method
they choose.